Fix #14 : Add integration with `project.version`
I can't say as I like the way I've implemented this: the conditionals
in the manifest metadata determinations are turning into quite the
non-OO tree.

That said, the contract test is solid, the code is still reasonably
readable, and I'm going to say that the feature addition is worth the
tradeoff.

Also, refactoring out the conditional tree feels like an exercise best
done on its own.

The Defense rests.  Committing.