Fix #15 : Add specific open source release license
Since Gradle itself is using the ASLv2 license, and Groovy has moved over
to the Apache Foundation (presumably adopting the ASLv2), I'm going to roll
with that for the time being. Not to be flip, but aligning ourselves with
the licensing for our parents and parters seems to me to lower the barrier
of entry for folks.