Remove logic from plugin
This is just a proposal. I think ideally the plugin would have little or
no logic—like a thin controller.

Some of the logic that was here I think is unnecessary. For example:

- `Version.warningText` is not valid code
- Explicitly declaring a project.version that's different from the source
  control tag is weird if you're opting into this plugin
- project.manifesto? seem necessary. If the project.manifesto extension
  doesn't exist, this plugin is fatally broken.

I think there is value in wrapping logic around attributes that might be
null, but maybe we should introduce another object to contain the logic that can
be tested in isolation.

Thoughts?