Fix regex for testing valid returns from describe
I'm not sure if we want all elements of a describe proper (the
equivalent to 'git describe' or 'git describe --tags') to be valid
for version numbers we'll accept, but this test is concerned with
getting a valid output from describe.

The new regex more accurately tests all possible describe results.